FIX: eeg2ftrip that works with strange eeglab behavior when some fields contain float 32 or int instead of double

+++ util/eeg2ftrip.m
@@ -67,8 +67,8 @@ end;
         
 switch fieldbox
   case 'preprocessing'
-      sampt = 1000/EEG.srate;
-      samptimes = EEG.times/sampt;
+      sampt = double(1000 / EEG.srate);
+      samptimes = double(EEG.times) / sampt;
       data.sampleinfo = zeros(EEG.trials,2);
       lastsamp = 0;
